Commandeering the 'bmw.py' name for the general module
