--- python/credentials.json
@@ -1,14 +1,0 @@
-{
-    "username": "whatever%40example.com", 
-    "password": "ABC*123",
-    "Authorization_Basic": "[BASE64 Encoded username:password decompiled from apk]",
-    "access_token": "",
-    "updateTime": "2015-11-20T23:35:18+0100", 
-    "twitter_access_token_secret": "[FROM TWITTER]", 
-    "twitter_consumer_secret": "[FROM TWITTER]", 
-    "twitter_access_token": "[FROM TWITTER]", 
-    "twitter_consumer_key": "[FROM TWITTER]",
-    "time": [
-        1448564884.470381
-    ]
-}
