+++ python/credentials.json.sample
@@ -1,0 +1,7 @@
+{
+    "username": "user@example.com", 
+    "auth_basic": "12345kNxdHhKdVhXUDc0eGYzQ0p3VUVQOjF6REh4NnVuNGNEanliTEVOTjNreWZ1bVgya0VZaWdXUGNRcGR2RFJwSUJrN3JPSg==", 
+    "password": "mypassword", 
+    "access_token": "XCudfgAhUpPtyuJIpENsk3Ec5y5HzC0V", 
+    "token_expiry": 1448773511.053424
+}
